Brute force fix for no init block
When a template is setup and has no init block, the init_done_flag_ is
forced true.

This places the template in a sane state.

However, with no init block at the top of the template, the entire
template is rendered.

So for large templates is suggested to specifiy an init block